requirements.txt
torch
numpy
matplotlib
openai
python-dotenv